+++ RELEASES.md
@@ -181,6 +181,7 @@ Upcoming
 - Better integration of Maven and Eclipse (using Kepler SR2).
 - added "Transaction Service" to console function "Show Security Items", and added configuration warning "receiver/pipeline transaction timeout exceeds system transaction timeout"
 - Improve Larva txt diff performance (especially for large messages)
+- Add fromClasspath to Larva XsltProviderListener